Backend/requirements.txt
# Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
# AWS IoT Core dependencies
boto3==1.34.0
awsiotsdk==1.21.0
paho-mqtt==1.6.1

# Twilio for WhatsApp notifications
twilio==8.10.0

# Twilio for WhatsApp/SMS notifications
twilio==8.10.0

# Machine Learning dependencies
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4

# Export dependencies
openpyxl==3.1.5
reportlab==4.4.4

# System monitoring
psutil==7.1.3

# Caching
redis==7.0.1
aioredis==2.0.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # For TestClient
faker==20.1.0  # For generating test data